Color picker view now recognizes when setColor: message is sent with a non-rgb colorspace model. In the case of CMYK color, the color is converted to RGB space.
This only solves part of the CMYK issue on the input end. That is, you can provide the picker with a CMYK color, but the color it sends back will still be in RBG colorspace.

TODO: The resulting color may not be appropriately scaled due to K normalization. Float values for the individual channels may end up being outside acceptable range. If [UIColor colorWithRed:Green:Blue] clamps the values, then things might be fine.